Add MatchMysqlDb class - mysql database backend
Take note that because of nonexisting data validation in view it was
necessary to implement it inside MatchMysqlDb.

As a result of that, MatchMysqlDb can and WILL raise exceptions when fed
wrong data. You should either handle this behaviour in your code or
rewrite view so it does validation and rewrite MatchMysqlDb so it
doesn't raise exceptions.